Pluralize the :topic initialization option
By default, kafka and kafka-go support consumers subscribing to
multiple topics, by passing an array of strings.

Rafka also respects this (see server.go#parseTopicsAndConfig).

Rafka-rb simply needs to change the wording in order to respect
the plural and in effect show the api user that consuming from
multiple topics is supported.

The commit takes the extra step of supporting an array of strings
as well as a single string, whether it contains a single topic or
multiple comma-separated topics.

Closes #5